requirements.txt
transformers==4.40.1
tensorflow==2.16.1
tensorflow==2.16.1
uvicorn[standard]
tf-keras
python-multipart==0.0.9
python-jose[cryptography]
passlib[bcrypt]
Jinja2==3.1.3
sqlmodel==0.0.18
alembic==1.13.1
pysqlite3-binary==0.5.2.post3
pytest==8.2.0
pytest-cov==5.0.0
pylint==3.2.0
black==24.4.2
ipython==8.24.0
openai==1.30.1
python-dotenv==1.0.1
pandas==2.2.2
